# docs/openapi.yaml
openapi: 3.1.0
info:
  title: account_Manager API
  version: 0.1.0
  description: Contrato versionado dos endpoints implementados até H02.3.
servers:
  - url: /api/v1
paths:
...
  /categories:
    get:
      operationId: listCategories
      summary: Lista categorias do espaço ativo
      security: [{ sessionCookie: [] }]
      parameters:
        - name: includeArchived
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Categorias do espaço, ordenadas por nome.
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Category" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createCategory
      summary: Cria uma categoria no espaço ativo
      security: [{ sessionCookie: [], xsrfCookie: [], xsrfHeader: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryNameRequest" }
      responses:
        "201":
          description: Categoria criada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { description: Nome já utilizado no espaço. }
  /categories/{id}:
    put:
      operationId: renameCategory
      summary: Renomeia uma categoria com controle otimista
      security: [{ sessionCookie: [], xsrfCookie: [], xsrfHeader: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RenameCategoryRequest" }
      responses:
        "200":
          description: Categoria renomeada; despesas vinculadas passam a exibir o novo nome.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Categoria não encontrada no espaço ativo. }
        "409": { description: Versão desatualizada, categoria arquivada ou nome duplicado. }
  /categories/{id}/archive:
    post:
      operationId: archiveCategory
      summary: Impede novas associações sem apagar referências existentes
      security: [{ sessionCookie: [], xsrfCookie: [], xsrfHeader: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryVersionRequest" }
      responses:
        "200":
          description: Categoria arquivada; histórico e despesas existentes preservados.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Categoria não encontrada no espaço ativo. }
        "409": { description: Versão desatualizada ou categoria já arquivada. }
  /expenses:
    post:
      operationId: createOneOffExpense
      summary: Cadastra uma despesa avulsa pendente ou já paga
      security:
        - sessionCookie: []
          xsrfCookie: []
          xsrfHeader: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: UUID preservado ao repetir a mesma intenção após falha de resposta.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseRequest"
      responses:
        "201":
          description: Despesa criada; Location identifica o recurso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "200":
          description: Repetição da mesma chave e payload; retorna a despesa original sem duplicar.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: A chave já foi utilizada com outro payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    get:
      operationId: listExpenses
      summary: Lista somente as despesas do espaço ativo do usuário
      security:
        - sessionCookie: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: sort
          in: query
          schema:
            type: string
            enum: [REFERENCE_DATE, AMOUNT, DESCRIPTION]
            default: REFERENCE_DATE
        - name: direction
          in: query
          schema:
            type: string
            enum: [ASC, DESC]
            default: ASC
      responses:
        "200":
          description: Página com ordenação estável e atraso calculado no fuso do espaço.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpensePage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /expenses/{id}:
    get:
      operationId: getExpense
      summary: Consulta uma despesa do espaço ativo para edição
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Despesa atual, incluindo sua versão otimista.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Expense" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: Despesa não encontrada no espaço ativo.
    put:
      operationId: correctExpense
      summary: Corrige os campos editáveis sem alterar a situação da despesa
      security:
        - sessionCookie: []
          xsrfCookie: []
          xsrfHeader: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: Idempotency-Key
          in: header
          required: true
          description: UUID preservado ao repetir a mesma correção após falha de resposta.
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CorrectExpenseRequest" }
      responses:
        "200":
          description: Correção aplicada ou repetição idempotente do mesmo resultado.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Expense" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: Despesa não encontrada no espaço ativo.
        "409":
          description: Versão ou situação desatualizada, ou chave reutilizada com outro payload. Consulte novamente antes de reaplicar; não há sobrescrita forçada.
  /expenses/{id}/payment:
    post:
      operationId: settleExpense
      summary: Quita integralmente uma despesa pendente
      security:
        - sessionCookie: []
          xsrfCookie: []
          xsrfHeader: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRequest"
      responses:
        "200":
          description: Quitação criada ou repetição idempotente do mesmo resultado.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Expense" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: Despesa não encontrada no espaço ativo.
        "409":
          description: Despesa já paga, versão desatualizada ou chave reutilizada com outro payload.
  /expenses/batch-payment:
    post:
      operationId: settleExpensesBatch
      summary: Quita integralmente vários lançamentos em uma única transação
      description: >-
        Data e pagador são comuns ao lote; cada item usa o valor confirmado da própria cobrança.
        Todos os itens são validados e bloqueados antes da gravação. Qualquer item indisponível,
        inelegível ou desatualizado rejeita o lote inteiro, sem efeitos parciais.
      security:
        - sessionCookie: []
          xsrfCookie: []
          xsrfHeader: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: UUID da intenção completa; deve ser preservado somente ao repetir o mesmo conteúdo.
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BatchPaymentRequest" }
      responses:
        "200":
          description: Lote aplicado integralmente ou replay idempotente do resultado original.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BatchPaymentResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: >-
            Lote integralmente rejeitado por item indisponível, não pendente, com valor não confirmado
            ou versão desatualizada; também cobre chave idempotente reutilizada com conteúdo diferente.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /expenses/{id}/payment-reversal:
    post:
      operationId: reverseExpensePayment
      summary: Desfaz a quitação ativa preservando pagamento e reversão no histórico
      security:
        - sessionCookie: []
          xsrfCookie: []
          xsrfHeader: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExpenseActionRequest" }
      responses:
        "200":
          description: Quitação desfeita ou repetição idempotente; a despesa volta a PENDING.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Expense" }
        "400":
          description: Motivo inválido ou despesa paga sem vencimento; nesse caso, informe o vencimento pelo fluxo de correção antes da reversão.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Despesa não encontrada no espaço ativo. }
        "409": { description: Situação ou versão desatualizada, ou chave reutilizada com outro payload. }
  /expenses/{id}/cancellation:
    post:
      operationId: cancelExpense
      summary: Cancela uma despesa pendente sem exclusão física
      security:
        - sessionCookie: []
          xsrfCookie: []
          xsrfHeader: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExpenseActionRequest" }
      responses:
        "200":
          description: Despesa cancelada ou repetição idempotente; permanece consultável por id e some da lista ativa.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Expense" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Despesa não encontrada no espaço ativo. }
        "409": { description: Despesa paga/cancelada, versão desatualizada ou chave reutilizada com outro payload. }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
    xsrfCookie:
      type: apiKey
      in: cookie
      name: XSRF-TOKEN
    xsrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
    setupSecret:
      type: apiKey
      in: header
      name: X-Setup-Secret
  responses:
    BadRequest:
      description: Campos ou invariantes inválidos.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthorized:
      description: Sessão ausente/expirada ou credenciais inválidas, conforme o endpoint.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Forbidden:
      description: Principal autenticado sem associação ativa ou sem o papel exigido no espaço.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
    Category:
      type: object
      required: [id, name, archived, version, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 60 }
        archived: { type: boolean }
        version: { type: integer, format: int64, minimum: 0 }
        updatedAt: { type: string, format: date-time }
    CategoryNameRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 60 }
    RenameCategoryRequest:
      allOf:
        - $ref: "#/components/schemas/CategoryNameRequest"
        - type: object
          required: [version]
          properties:
            version: { type: integer, format: int64, minimum: 0 }
    CategoryVersionRequest:
      type: object
      additionalProperties: false
      required: [version]
      properties:
        version: { type: integer, format: int64, minimum: 0 }
    CreateExpenseRequest:
      type: object
      additionalProperties: false
      required: [description, amount, status]
      properties:
        description:
          type: string
          minLength: 1
          maxLength: 200
        amount:
          type: string
          pattern: '^\d{1,8}(\.\d{1,2})?$'
          example: "150.00"
        status:
          type: string
          enum: [PENDING, PAID]
        dueDate:
          type: [string, "null"]
          format: date
          description: Obrigatório para PENDING; opcional para PAID.
        paymentDate:
          type: [string, "null"]
          format: date
          description: Obrigatório para PAID e ausente para PENDING.
        notes:
          type: [string, "null"]
          maxLength: 2000
        paidAmount:
          type: [string, "null"]
          pattern: '^\d{1,8}(\.\d{1,2})?$'
          description: Para PAID, valor efetivo; omissão sugere o valor da cobrança.
        paidByUserId:
          type: [string, "null"]
          format: uuid
          description: Para PAID, membro ativo pagador; omissão sugere o autor.
        paymentNotes:
          type: [string, "null"]
          maxLength: 2000
        categoryId:
          type: [string, "null"]
          format: uuid
          description: Categoria ativa do espaço; ausência representa Sem categoria.
    PaymentRequest:
      type: object
      additionalProperties: false
      required: [version, paidAmount, paymentDate, paidByUserId]
      properties:
        version: { type: integer, format: int64, minimum: 0 }
        paidAmount:
          type: string
          pattern: '^\d{1,8}(\.\d{1,2})?$'
        paymentDate: { type: string, format: date }
        paidByUserId: { type: string, format: uuid }
        paymentNotes: { type: [string, "null"], maxLength: 2000 }
        categoryId:
          type: [string, "null"]
          format: uuid
          description: Categoria ativa do espaço, ou a categoria arquivada já vinculada quando não houver troca.
    BatchPaymentRequest:
      type: object
      additionalProperties: false
      required: [items, paymentDate, paidByUserId, confirmed]
      properties:
        items:
          type: array
          minItems: 1
          description: Seleção sem IDs repetidos; a ordem não altera a identidade da intenção.
          items:
            type: object
            additionalProperties: false
            required: [expenseId, version]
            properties:
              expenseId: { type: string, format: uuid }
              version: { type: integer, format: int64, minimum: 0 }
        paymentDate: { type: string, format: date }
        paidByUserId:
          type: string
          format: uuid
          description: Membro ativo do mesmo espaço, comum a todos os itens.
        confirmed:
          const: true
          description: Confirmação explícita da quitação integral de todos ou nenhum.
    BatchPaymentResult:
      type: object
      required: [operationId, items, replayed]
      properties:
        operationId: { type: string, format: uuid }
        replayed: { type: boolean }
        items:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [expenseId, fromVersion, toVersion, paidAmount]
            properties:
              expenseId: { type: string, format: uuid }
              fromVersion: { type: integer, format: int64, minimum: 0 }
              toVersion: { type: integer, format: int64, minimum: 1 }
              paidAmount: { type: string, pattern: '^\d+\.\d{2}$' }
    CorrectExpenseRequest:
      type: object
      additionalProperties: false
      required: [version, status, description, amount]
      properties:
        version: { type: integer, format: int64, minimum: 0 }
        status:
          type: string
          enum: [PENDING, PAID]
          description: Deve coincidir com a situação carregada; correção não muda o estado.
        description: { type: string, minLength: 1, maxLength: 200 }
        amount:
          type: string
          pattern: '^\d{1,8}(\.\d{1,2})?$'
        dueDate:
          type: [string, "null"]
          format: date
          description: Obrigatório para PENDING e opcional para PAID.
        notes: { type: [string, "null"], maxLength: 2000 }
        paidAmount:
          type: [string, "null"]
          pattern: '^\d{1,8}(\.\d{1,2})?$'
          description: Obrigatório somente para PAID.
        paymentDate:
          type: [string, "null"]
          format: date
          description: Obrigatório somente para PAID.
        paidByUserId:
          type: [string, "null"]
          format: uuid
          description: Membro ativo pagador; obrigatório somente para PAID.
        paymentNotes: { type: [string, "null"], maxLength: 2000 }
    ExpenseActionRequest:
      type: object
      additionalProperties: false
      required: [version, reason]
      properties:
        version: { type: integer, format: int64, minimum: 0 }
        reason: { type: string, minLength: 1, maxLength: 2000 }
    Expense:
      type: object
      required: [id, origin, description, amount, currency, status, referenceDate, overdue,
        createdByUserId, createdByDisplayName, createdAt, version, history]
      properties:
        id: { type: string, format: uuid }
        origin: { const: ONE_OFF }
        description: { type: string }
        amount:
          type: string
          pattern: '^\d+\.\d{2}$'
        currency: { const: BRL }
        status: { type: string, enum: [PENDING, PAID, CANCELLED] }
        dueDate: { type: [string, "null"], format: date }
        paymentDate: { type: [string, "null"], format: date }
        paidAmount:
          type: [string, "null"]
          pattern: '^\d+\.\d{2}$'
        referenceDate: { type: string, format: date }
        overdue:
          type: boolean
          description: Somente PENDING com vencimento anterior à data local do espaço.
        categoryId: { type: [string, "null"], format: uuid }
        categoryName:
          type: [string, "null"]
          description: Nome atual preservado para leitura; nulo é apresentado como Sem categoria.
        responsibleUserId: { type: [string, "null"], format: uuid }
        notes: { type: [string, "null"] }
        createdByUserId: { type: string, format: uuid }
        createdByDisplayName: { type: string }
        paidByUserId: { type: [string, "null"], format: uuid }
        paidByDisplayName: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
        version: { type: integer, format: int64, minimum: 0 }
        paymentAudit:
          oneOf:
            - $ref: "#/components/schemas/PaymentAudit"
            - type: "null"
        history:
          type: array
          description: Eventos históricos no detalhe; a listagem retorna lista vazia.
          items: { $ref: "#/components/schemas/ExpenseHistoryEvent" }
    PaymentAudit:
      type: object
      required: [recordedByUserId, recordedByDisplayName, recordedAt]
      properties:
        recordedByUserId: { type: string, format: uuid }
        recordedByDisplayName: { type: string }
        recordedAt: { type: string, format: date-time }
        notes: { type: [string, "null"], maxLength: 2000 }
    ExpenseHistoryEvent:
      type: object
      required: [type, actorUserId, actorDisplayName, occurredAt, version]
      properties:
        type:
          type: string
          enum: [EXPENSE_PAID, PAYMENT_REVERSED, EXPENSE_CORRECTED, EXPENSE_CANCELLED]
        actorUserId: { type: string, format: uuid }
        actorDisplayName: { type: string }
        occurredAt: { type: string, format: date-time }
        reason: { type: [string, "null"], maxLength: 2000 }
        notes: { type: [string, "null"], maxLength: 2000 }
        version: { type: integer, format: int64, minimum: 0 }
        paidAmount: { type: [string, "null"], pattern: '^\d+\.\d{2}$' }
        paymentDate: { type: [string, "null"], format: date }
        payerUserId: { type: [string, "null"], format: uuid }
        payerDisplayName: { type: [string, "null"] }
        changedFields: { type: [string, "null"] }
        batchOperationId:
          type: [string, "null"]
          format: uuid
          description: Correlação comum às auditorias individuais produzidas por uma quitação em lote.
    ExpensePage:
      type: object
      required: [content, page, size, totalElements, totalPages, sort, direction]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Expense"
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1, maximum: 100 }
        totalElements: { type: integer, format: int64, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
        sort: { type: string, enum: [REFERENCE_DATE, AMOUNT, DESCRIPTION] }
        direction: { type: string, enum: [ASC, DESC] }
...
    ApiError:
      type: object
      required: [code, message, fieldErrors, operationId]
      properties:
        code:
          type: string
        message:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        operationId:
          type: string
          format: uuid
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string